+++ BaseOAuth.php
@@ -10,7 +10,6 @@ namespace yii\authclient;
 use yii\base\Exception;
 use yii\base\InvalidParamException;
 use Yii;
-use yii\helpers\Json;
 
 /**
  * BaseOAuth is a base class for the OAuth clients.
@@ -19,7 +18,6 @@ use yii\helpers\Json;
  *
  * @property OAuthToken $accessToken Auth token instance. Note that the type of this property differs in
  * getter and setter. See [[getAccessToken()]] and [[setAccessToken()]] for details.
- * @property array $curlOptions CURL options. This property is read-only.
  * @property string $returnUrl Return URL.
  * @property signature\BaseMethod $signatureMethod Signature method instance. Note that the type of this
  * property differs in getter and setter. See [[getSignatureMethod()]] and [[setSignatureMethod()]] for details.
@@ -27,19 +25,16 @@ use yii\helpers\Json;
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0
  */
-abstract class BaseOAuth extends BaseClient implements ClientInterface
+abstract class BaseOAuth extends BaseClient
 {
-    const CONTENT_TYPE_JSON = 'json'; // JSON format
-    const CONTENT_TYPE_URLENCODED = 'urlencoded'; // urlencoded query string, like name1=value1&name2=value2
-    const CONTENT_TYPE_XML = 'xml'; // XML format
-    const CONTENT_TYPE_AUTO = 'auto'; // attempts to determine format automatically
-
     /**
      * @var string protocol version.
      */
     public $version = '1.0';
     /**
      * @var string API base URL.
+     * This field will be used as [[\yii\httpclient\Client::baseUrl]] value of [[httpClient]].
+     * Note: changing this property will take no effect after [[httpClient]] is instantiated.
      */
     public $apiBaseUrl;
     /**
@@ -62,11 +57,6 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
      * By default current URL will be used.
      */
     private $_returnUrl;
-    /**
-     * @var array cURL request options. Option values from this field will overwrite corresponding
-     * values from [[defaultCurlOptions()]].
-     */
-    private $_curlOptions = [];
     /**
      * @var OAuthToken|array access token instance or its array configuration.
      */
@@ -97,23 +87,8 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
-     * @param array $curlOptions cURL options.
-     */
-    public function setCurlOptions(array $curlOptions)
-    {
-        $this->_curlOptions = $curlOptions;
-    }
-
-    /**
-     * @return array cURL options.
-     */
-    public function getCurlOptions()
-    {
-        return $this->_curlOptions;
-    }
-
-    /**
-     * @param array|OAuthToken $token
+     * Sets access token to be used.
+     * @param array|OAuthToken $token access token or its configuration.
      */
     public function setAccessToken($token)
     {
@@ -137,6 +112,7 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
+     * Set signature method to be used.
      * @param array|signature\BaseMethod $signatureMethod signature method instance or its array configuration.
      * @throws InvalidParamException on wrong argument.
      */
@@ -161,6 +137,28 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
+     * @inheritdoc
+     */
+    public function setHttpClient($httpClient)
+    {
+        if (is_object($httpClient)) {
+            $httpClient = clone $httpClient;
+            $httpClient->baseUrl = $this->apiBaseUrl;
+        }
+        parent::setHttpClient($httpClient);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function createHttpClient($reference)
+    {
+        $httpClient = parent::createHttpClient($reference);
+        $httpClient->baseUrl = $this->apiBaseUrl;
+        return $httpClient;
+    }
+
+    /**
      * Composes default [[returnUrl]] value.
      * @return string return URL.
      */
@@ -170,187 +168,15 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
-     * Sends HTTP request.
-     * @param string $method request type.
-     * @param string $url request URL.
-     * @param array $params request params.
-     * @param array $headers additional request headers.
-     * @return array response.
-     * @throws Exception on failure.
+     * @inheritdoc
      */
-    protected function sendRequest($method, $url, array $params = [], array $headers = [])
-    {
-        $curlOptions = $this->mergeCurlOptions(
-            $this->defaultCurlOptions(),
-            $this->getCurlOptions(),
-            [
-                CURLOPT_HTTPHEADER => $headers,
-                CURLOPT_RETURNTRANSFER => true,
-                CURLOPT_URL => $url,
-            ],
-            $this->composeRequestCurlOptions(strtoupper($method), $url, $params)
-        );
-        $curlResource = curl_init();
-        foreach ($curlOptions as $option => $value) {
-            curl_setopt($curlResource, $option, $value);
-        }
-        $response = curl_exec($curlResource);
-        $responseHeaders = curl_getinfo($curlResource);
-
-        // check cURL error
-        $errorNumber = curl_errno($curlResource);
-        $errorMessage = curl_error($curlResource);
-
-        curl_close($curlResource);
-
-        if ($errorNumber > 0) {
-            throw new Exception('Curl error requesting "' .  $url . '": #' . $errorNumber . ' - ' . $errorMessage);
-        }
-        if (strncmp($responseHeaders['http_code'], '20', 2) !== 0) {
-            throw new InvalidResponseException($responseHeaders, $response, 'Request failed with code: ' . $responseHeaders['http_code'] . ', message: ' . $response);
-        }
-
-        return $this->processResponse($response, $this->determineContentTypeByHeaders($responseHeaders));
-    }
-
-    /**
-     * Merge CUrl options.
-     * If each options array has an element with the same key value, the latter
-     * will overwrite the former.
-     * @param array $options1 options to be merged to.
-     * @param array $options2 options to be merged from. You can specify additional
-     * arrays via third argument, fourth argument etc.
-     * @return array merged options (the original options are not changed.)
-     */
-    protected function mergeCurlOptions($options1, $options2)
-    {
-        $args = func_get_args();
-        $res = array_shift($args);
-        while (!empty($args)) {
-            $next = array_shift($args);
-            foreach ($next as $k => $v) {
-                if (is_array($v) && !empty($res[$k]) && is_array($res[$k])) {
-                    $res[$k] = array_merge($res[$k], $v);
-                } else {
-                    $res[$k] = $v;
-                }
-            }
-        }
-        return $res;
-    }
-
-    /**
-     * Returns default cURL options.
-     * @return array cURL options.
-     */
-    protected function defaultCurlOptions()
+    protected function defaultRequestOptions()
     {
         return [
-            CURLOPT_USERAGENT => Yii::$app->name . ' OAuth ' . $this->version . ' Client',
-            CURLOPT_CONNECTTIMEOUT => 30,
-            CURLOPT_TIMEOUT => 30,
-            CURLOPT_SSL_VERIFYPEER => false,
+            'userAgent' => Yii::$app->name . ' OAuth ' . $this->version . ' Client',
+            'timeout' => 30,
+            'sslVerifyPeer' => false,
         ];
-    }
-
-    /**
-     * Processes raw response converting it to actual data.
-     * @param string $rawResponse raw response.
-     * @param string $contentType response content type.
-     * @throws Exception on failure.
-     * @return array actual response.
-     */
-    protected function processResponse($rawResponse, $contentType = self::CONTENT_TYPE_AUTO)
-    {
-        if (empty($rawResponse)) {
-            return [];
-        }
-        switch ($contentType) {
-            case self::CONTENT_TYPE_AUTO: {
-                $contentType = $this->determineContentTypeByRaw($rawResponse);
-                if ($contentType == self::CONTENT_TYPE_AUTO) {
-                    throw new Exception('Unable to determine response content type automatically.');
-                }
-                $response = $this->processResponse($rawResponse, $contentType);
-                break;
-            }
-            case self::CONTENT_TYPE_JSON: {
-                $response = Json::decode($rawResponse, true);
-                break;
-            }
-            case self::CONTENT_TYPE_URLENCODED: {
-                $response = [];
-                parse_str($rawResponse, $response);
-                break;
-            }
-            case self::CONTENT_TYPE_XML: {
-                $response = $this->convertXmlToArray($rawResponse);
-                break;
-            }
-            default: {
-                throw new Exception('Unknown response type "' . $contentType . '".');
-            }
-        }
-        return $response;
-    }
-
-    /**
-     * Converts XML document to array.
-     * @param string|\SimpleXMLElement $xml xml to process.
-     * @return array XML array representation.
-     */
-    protected function convertXmlToArray($xml)
-    {
-        if (!is_object($xml)) {
-            $xml = simplexml_load_string($xml);
-        }
-        $result = (array) $xml;
-        foreach ($result as $key => $value) {
-            if (is_object($value)) {
-                $result[$key] = $this->convertXmlToArray($value);
-            }
-        }
-        return $result;
-    }
-
-    /**
-     * Attempts to determine HTTP request content type by headers.
-     * @param array $headers request headers.
-     * @return string content type.
-     */
-    protected function determineContentTypeByHeaders(array $headers)
-    {
-        if (isset($headers['content_type'])) {
-            if (stripos($headers['content_type'], 'json') !== false) {
-                return self::CONTENT_TYPE_JSON;
-            }
-            if (stripos($headers['content_type'], 'urlencoded') !== false) {
-                return self::CONTENT_TYPE_URLENCODED;
-            }
-            if (stripos($headers['content_type'], 'xml') !== false) {
-                return self::CONTENT_TYPE_XML;
-            }
-        }
-        return self::CONTENT_TYPE_AUTO;
-    }
-
-    /**
-     * Attempts to determine the content type from raw content.
-     * @param string $rawContent raw response content.
-     * @return string response type.
-     */
-    protected function determineContentTypeByRaw($rawContent)
-    {
-        if (preg_match('/^\\{.*\\}$/is', $rawContent)) {
-            return self::CONTENT_TYPE_JSON;
-        }
-        if (preg_match('/^[^=|^&]+=[^=|^&]+(&[^=|^&]+=[^=|^&]+)*$/is', $rawContent)) {
-            return self::CONTENT_TYPE_URLENCODED;
-        }
-        if (preg_match('/^<.*>$/is', $rawContent)) {
-            return self::CONTENT_TYPE_XML;
-        }
-        return self::CONTENT_TYPE_AUTO;
     }
 
     /**
@@ -380,6 +206,24 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
+     * Sends the given HTTP request, returning response data.
+     * @param \yii\httpclient\Request $request HTTP request to be sent.
+     * @return array response data.
+     * @throws InvalidResponseException on invalid remote response.
+     * @since 2.1
+     */
+    protected function sendRequest($request)
+    {
+        $response = $request->send();
+
+        if (!$response->getIsOk()) {
+            throw new InvalidResponseException($response, 'Request failed with code: ' . $response->getStatusCode() . ', message: ' . $response->getContent());
+        }
+
+        return $response->getData();
+    }
+
+    /**
      * Composes URL from base URL and GET params.
      * @param string $url base URL.
      * @param array $params GET params.
@@ -387,12 +231,14 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
      */
     protected function composeUrl($url, array $params = [])
     {
-        if (strpos($url, '?') === false) {
-            $url .= '?';
-        } else {
-            $url .= '&';
+        if (!empty($params)) {
+            if (strpos($url, '?') === false) {
+                $url .= '?';
+            } else {
+                $url .= '&';
+            }
+            $url .= http_build_query($params, '', '&', PHP_QUERY_RFC3986);
         }
-        $url .= http_build_query($params, '', '&', PHP_QUERY_RFC3986);
         return $url;
     }
 
@@ -484,37 +330,52 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     }
 
     /**
-     * Performs request to the OAuth API.
-     * @param string $apiSubUrl API sub URL, which will be append to [[apiBaseUrl]], or absolute API URL.
-     * @param string $method request method.
-     * @param array $params request parameters.
-     * @param array $headers additional request headers.
-     * @return array API response
+     * Creates an HTTP request for the API call.
+     * @see createRequest()
+     * @param array $config request object configuration.
+     * @return \yii\httpclient\Request request instance.
      * @throws Exception on failure.
+     * @since 2.1
      */
-    public function api($apiSubUrl, $method = 'GET', array $params = [], array $headers = [])
+    public function createApiRequest(array $config = [])
     {
-        if (preg_match('/^https?:\\/\\//is', $apiSubUrl)) {
-            $url = $apiSubUrl;
+        if (isset($config['accessToken'])) {
+            $accessToken = $config['accessToken'];
         } else {
-            $url = $this->apiBaseUrl . '/' . $apiSubUrl;
+            $accessToken = $this->getAccessToken();
         }
-        $accessToken = $this->getAccessToken();
         if (!is_object($accessToken) || !$accessToken->getIsValid()) {
             throw new Exception('Invalid access token.');
         }
-        return $this->apiInternal($accessToken, $url, $method, $params, $headers);
+
+        $request = $this->createRequest($config);
+
+        $this->applyAccessTokenToRequest($request, $accessToken);
+
+        return $request;
     }
 
     /**
-     * Composes HTTP request CUrl options, which will be merged with the default ones.
-     * @param string $method request type.
-     * @param string $url request URL.
-     * @param array $params request params.
-     * @return array CUrl options.
-     * @throws Exception on failure.
+     * Performs request to the OAuth API returning response data.
+     * You may use [[createApiRequest()]] method instead gaining more control over request execution.
+     * @see createApiRequest()
+     * @param string $apiSubUrl API sub URL, which will be append to [[apiBaseUrl]], or absolute API URL.
+     * @param string $method request method.
+     * @param array|string $data request data or content.
+     * @param array $headers additional request headers.
+     * @return array API response data.
      */
-    abstract protected function composeRequestCurlOptions($method, $url, array $params);
+    public function api($apiSubUrl, $method = 'GET', $data = [], $headers = [])
+    {
+        $request = $this->createApiRequest([
+            'url' => $apiSubUrl,
+            'method' => $method,
+            'data' => $data,
+            'headers' => $headers,
+        ]);
+
+        return $this->sendRequest($request);
+    }
 
     /**
      * Gets new auth token to replace expired one.
@@ -524,14 +385,10 @@ abstract class BaseOAuth extends BaseClient implements ClientInterface
     abstract public function refreshAccessToken(OAuthToken $token);
 
     /**
-     * Performs request to the OAuth API.
-     * @param OAuthToken $accessToken actual access token.
-     * @param string $url absolute API URL.
-     * @param string $method request method.
-     * @param array $params request parameters.
-     * @param array $headers additional request headers.
-     * @return array API response.
-     * @throws Exception on failure.
+     * Applies access token to the HTTP request instance.
+     * @param \yii\httpclient\Request $request HTTP request instance.
+     * @param OAuthToken $accessToken access token instance.
+     * @since 2.1
      */
-    abstract protected function apiInternal($accessToken, $url, $method, array $params, array $headers);
+    abstract public function applyAccessTokenToRequest($request, $accessToken);
 }