+++ Collection.php
@@ -22,7 +22,7 @@ use Yii;
  *         'class' => 'yii\authclient\Collection',
  *         'clients' => [
  *             'google' => [
- *                 'class' => 'yii\authclient\clients\GoogleOAuth',
+ *                 'class' => 'yii\authclient\clients\Google',
  *                 'clientId' => 'google_client_id',
  *                 'clientSecret' => 'google_client_secret',
  *              ],
@@ -44,6 +44,13 @@ use Yii;
  */
 class Collection extends Component
 {
+    /**
+     * @var \yii\httpclient\Client|array|string HTTP client instance or configuration for the [[clients]].
+     * If set, this value will be passed as 'httpClient' config option while instantiating particular client object.
+     * This option is useful for adjusting HTTP client configuration for the entire list of auth clients.
+     */
+    public $httpClient;
+
     /**
      * @var array list of Auth clients with their configuration in format: 'clientId' => [...]
      */
@@ -107,6 +114,10 @@ class Collection extends Component
     protected function createClient($id, $config)
     {
         $config['id'] = $id;
+
+        if (!isset($config['httpClient']) && $this->httpClient !== null) {
+            $config['httpClient'] = $this->httpClient;
+        }
 
         return Yii::createObject($config);
     }