+++ tests/BaseClientTest.php
@@ -8,7 +8,7 @@ class BaseClientTest extends TestCase
 {
     public function testSetGet()
     {
-        $client = new Client();
+        $client = new ClientMock();
 
         $id = 'test_id';
         $client->setId($id);
@@ -42,11 +42,18 @@ class BaseClientTest extends TestCase
         ];
         $client->setViewOptions($viewOptions);
         $this->assertEquals($viewOptions, $client->getViewOptions(), 'Unable to setup view options!');
+
+        $requestOptions = [
+            'option1' => 'value1',
+            'option2' => 'value2',
+        ];
+        $client->setRequestOptions($requestOptions);
+        $this->assertEquals($requestOptions, $client->getRequestOptions(), 'Unable to setup request options!');
     }
 
     public function testGetDefaults()
     {
-        $client = new Client();
+        $client = new ClientMock();
 
         $this->assertNotEmpty($client->getName(), 'Unable to get default name!');
         $this->assertNotEmpty($client->getTitle(), 'Unable to get default title!');
@@ -141,7 +148,7 @@ class BaseClientTest extends TestCase
      */
     public function testNormalizeUserAttributes($normalizeUserAttributeMap, $rawUserAttributes, $expectedNormalizedUserAttributes)
     {
-        $client = new Client();
+        $client = new ClientMock();
         $client->setNormalizeUserAttributeMap($normalizeUserAttributeMap);
 
         $client->setUserAttributes($rawUserAttributes);
@@ -149,8 +156,25 @@ class BaseClientTest extends TestCase
 
         $this->assertEquals(array_merge($rawUserAttributes, $expectedNormalizedUserAttributes), $normalizedUserAttributes);
     }
+
+    public function testSetupHttpClient()
+    {
+        $client = new ClientMock();
+
+        $client->setHttpClient([
+            'baseUrl' => 'http://domain.com'
+        ]);
+        $httpClient = $client->getHttpClient();
+
+        $this->assertTrue($httpClient instanceof \yii\httpclient\Client, 'Unable to setup http client.');
+        $this->assertEquals('http://domain.com', $httpClient->baseUrl, 'Unable to setup http client property.');
+
+        $client = new ClientMock();
+        $httpClient = $client->getHttpClient();
+        $this->assertTrue($httpClient instanceof \yii\httpclient\Client, 'Unable to get default http client.');
+    }
 }
 
-class Client extends BaseClient
+class ClientMock extends BaseClient
 {
 }